requirements.txt
feedparser==6.0.10
keras==2.15.0
matplotlib==3.7.2
numpy==1.26.2
pandas==2.0.3
Pillow==9.2.0
Pillow==10.1.0
scikit_learn==1.3.0
streamlit==1.28.2
